changed back to the old BI events files and only added the new ones
