+++ core/src/main/java/com/kin/ecosystem/core/bi/events/Client.java
@@ -40,6 +40,14 @@ public class Client implements ClientInterface {
      * (Required)
      * 
      */
+    @SerializedName("device_id")
+    @Expose
+    private String deviceId;
+    /**
+     * 
+     * (Required)
+     * 
+     */
     @SerializedName("device_manufacturer")
     @Expose
     private String deviceManufacturer;
@@ -66,12 +74,14 @@ public class Client implements ClientInterface {
      * @param language
      * @param deviceModel
      * @param deviceManufacturer
+     * @param deviceId
      */
-    public Client(String os, String language, String carrier, String deviceManufacturer, String deviceModel) {
+    public Client(String os, String language, String carrier, String deviceId, String deviceManufacturer, String deviceModel) {
         super();
         this.os = os;
         this.language = language;
         this.carrier = carrier;
+        this.deviceId = deviceId;
         this.deviceManufacturer = deviceManufacturer;
         this.deviceModel = deviceModel;
     }
@@ -128,6 +138,24 @@ public class Client implements ClientInterface {
      */
     public void setCarrier(String carrier) {
         this.carrier = carrier;
+    }
+
+    /**
+     * 
+     * (Required)
+     * 
+     */
+    public String getDeviceId() {
+        return deviceId;
+    }
+
+    /**
+     * 
+     * (Required)
+     * 
+     */
+    public void setDeviceId(String deviceId) {
+        this.deviceId = deviceId;
     }
 
     /**
