+++ core/src/main/java/com/kin/ecosystem/core/bi/events/ClientInterface.java
@@ -7,6 +7,8 @@ public interface ClientInterface extends ClientReadonly {
 
     void setCarrier(String carrier);
 
+    void setDeviceId(String deviceId);
+
     void setDeviceManufacturer(String deviceManufacturer);
 
     void setDeviceModel(String deviceModel);
