+++ core/src/main/java/com/kin/ecosystem/core/bi/events/ClientReadonly.java
@@ -9,6 +9,8 @@ public interface ClientReadonly {
 
     String getCarrier();
 
+    String getDeviceId();
+
     String getDeviceManufacturer();
 
     String getDeviceModel();
