+++ core/src/main/java/com/kin/ecosystem/core/bi/events/CommonReadonly.java
@@ -9,8 +9,6 @@ public interface CommonReadonly {
 
     String getVersion();
 
-    String getDeviceId();
-
     String getUserId();
 
     Long getTimestamp();
