+++ core/src/main/java/com/kin/ecosystem/core/bi/events/UserLoginFailed.java
@@ -2,18 +2,15 @@
 package com.kin.ecosystem.core.bi.events;
 
 // Augmented by script
-import com.kin.ecosystem.core.bi.Event;
-import com.kin.ecosystem.core.bi.EventsStore;
-
-import java.util.HashMap;
-import java.util.Map;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+import com.kin.ecosystem.core.bi.Event;
+import com.kin.ecosystem.core.bi.EventsStore;
 
 
 /**
  * User login failed
- * 
+ *
  */
 public class UserLoginFailed implements Event {
     public static final String EVENT_NAME = "user_login_failed";
@@ -29,17 +26,17 @@ public class UserLoginFailed implements Event {
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     @SerializedName("event_name")
     @Expose
     private String eventName = EVENT_NAME;
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     @SerializedName("event_type")
     @Expose
@@ -47,7 +44,7 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all events
      * (Required)
-     * 
+     *
      */
     @SerializedName("common")
     @Expose
@@ -55,7 +52,7 @@ public class UserLoginFailed implements Event {
     /**
      * common user properties
      * (Required)
-     * 
+     *
      */
     @SerializedName("user")
     @Expose
@@ -63,15 +60,15 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all client events
      * (Required)
-     * 
+     *
      */
     @SerializedName("client")
     @Expose
     private Client client;
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     @SerializedName("error_reason")
     @Expose
@@ -79,13 +76,13 @@ public class UserLoginFailed implements Event {
 
     /**
      * No args constructor for use in serialization
-     * 
+     *
      */
     public UserLoginFailed() {
     }
 
     /**
-     * 
+     *
      * @param common
      * @param errorReason
 
@@ -102,36 +99,36 @@ public class UserLoginFailed implements Event {
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public String getEventName() {
         return eventName;
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public void setEventName(String eventName) {
         this.eventName = eventName;
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public String getEventType() {
         return eventType;
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public void setEventType(String eventType) {
         this.eventType = eventType;
@@ -140,7 +137,7 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all events
      * (Required)
-     * 
+     *
      */
     public Common getCommon() {
         return common;
@@ -149,7 +146,7 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all events
      * (Required)
-     * 
+     *
      */
     public void setCommon(Common common) {
         this.common = common;
@@ -158,7 +155,7 @@ public class UserLoginFailed implements Event {
     /**
      * common user properties
      * (Required)
-     * 
+     *
      */
     public User getUser() {
         return user;
@@ -167,7 +164,7 @@ public class UserLoginFailed implements Event {
     /**
      * common user properties
      * (Required)
-     * 
+     *
      */
     public void setUser(User user) {
         this.user = user;
@@ -176,7 +173,7 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all client events
      * (Required)
-     * 
+     *
      */
     public Client getClient() {
         return client;
@@ -185,25 +182,25 @@ public class UserLoginFailed implements Event {
     /**
      * common properties for all client events
      * (Required)
-     * 
+     *
      */
     public void setClient(Client client) {
         this.client = client;
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public String getErrorReason() {
         return errorReason;
     }
 
     /**
-     * 
+     *
      * (Required)
-     * 
+     *
      */
     public void setErrorReason(String errorReason) {
         this.errorReason = errorReason;
