+++ core/src/main/java/com/kin/ecosystem/core/bi/events/UserLoginSucceeded.java
@@ -2,13 +2,10 @@
 package com.kin.ecosystem.core.bi.events;
 
 // Augmented by script
-import com.kin.ecosystem.core.bi.Event;
-import com.kin.ecosystem.core.bi.EventsStore;
-
-import java.util.HashMap;
-import java.util.Map;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+import com.kin.ecosystem.core.bi.Event;
+import com.kin.ecosystem.core.bi.EventsStore;
 
 
 /**
@@ -78,9 +75,7 @@ public class UserLoginSucceeded implements Event {
     /**
      * 
      * @param common
-
      * @param client
-
      * @param user
      */
     public UserLoginSucceeded(Common common, User user, Client client) {
@@ -179,5 +174,4 @@ public class UserLoginSucceeded implements Event {
     public void setClient(Client client) {
         this.client = client;
     }
-
 }
