+++ test-base/src/main/java/kin/ecosystem/test/base/ClientModifierFake.kt
@@ -9,6 +9,7 @@ class ClientModifierFake : ClientModifier {
     private var os: String = "android"
     private var language: String = "english"
     private var carrier: String = "test_carrier"
+    private var deviceId: String = "test_device_id"
     private var deviceManufacturer: String = "test_manufacturer"
     private var deviceModel: String = "test_device_model"
 
@@ -16,6 +17,7 @@ class ClientModifierFake : ClientModifier {
         mutable.setOs { os }
         mutable.setLanguage { language }
         mutable.setCarrier { carrier }
+        mutable.setDeviceId { deviceId }
         mutable.setDeviceManufacturer { deviceManufacturer }
         mutable.setDeviceModel { deviceModel }
     }
