+++ test-base/src/main/java/kin/ecosystem/test/base/CommonModifierFake.kt
@@ -10,14 +10,12 @@ class CommonModifierFake : CommonModifier {
     private var eventId: UUID = UUID.randomUUID()
     private var version: String = "test_version"
     private var userId: String = "test_user_id"
-    private var deviceId: String = "test_device_id"
     private var timestamp: Long = System.currentTimeMillis()
 
     override fun modify(mutable: CommonProxy) {
         mutable.setEventId { eventId }
         mutable.setVersion { version }
         mutable.setUserId { userId }
-        mutable.setDeviceId { deviceId }
         mutable.setTimestamp { timestamp }
     }
 }
